:sparkles: True Flocking and fix warnings

+++ includes/Flock.hpp
@@ -11,7 +11,7 @@ public:
   Flock() = default;
   ~Flock();
 
-  void update(const double &dT);
+  void update(const float &dT);
   void render();
   void add_boid(const glm::vec3 &pos);
   const std::vector<Boid*> in_fov_of(const Boid &boid);
